libmsmcomm: some automake enhancements
I found that some things were missing in order to have a working "make
dist".
Here is a patch against libmsmcomm's latest rev (git.freesmartphone.org)
at the time of writing which creates a "MAINTAINERS" file and adds 3
headers in the list of the sources to be included in the tarball.

Thanks David 'Deubeuliou' for spotting this!