msmcommd: if we send more sync frames than expected reset the modem
Sometimes msmcommd comes in an infinite loop where it tries to send a sync frame to the
modem but the modem never responses with a sync resp frame. To avoid this we tell the
timer logic to allow only a certain number of times to send the sync frame to the modem.
If the number of maximum tries is reached the modem is reseted and link layer setup is
started again.

Signed-off-by: Simon Busch <morphis@gravedo.de>