msmcommd: fix modem/link reset handling
We have to differentiate between a modem and a link reset and when we have to use them. If
the link runs out of sync we reset only the link not the whole modem! If the link layer
detects that something is wrong with the modem (e.g. it never gets a response) than we
reset the modem and the link layer.

Signed-off-by: Simon Busch <morphis@gravedo.de>