libmsmcomm: return of length attribute in message structures
In some cases we need to access to length attribute of message structures. If we just copy
data with the Memory.copy method to the array it don't updates the length property which
is necessary to send a correct formated message to the modem. If there is some way to
update the length property automatically we can revert this.