msmcommd: simservice: work around casting response messages to early
When we cast the response message right after we got them we cannot be sure that they have
the correct message type. We have to check the message type first and the proceed with
parsing the response message data. If we do not the service will die cause it's not clear
when we get a ReturnResponseMessage and when a CallbackResponseMessage.

The way this is currently realize is not the best. If have to think about how to make it
better before I start to port it to the other services ...