msmcommd: refactor message response processing into callback functions
While working with the msmcomm protocol I discovered there are some situations where you
need to process different response types for one send message. For example you send a
command message and the result is in success case a unsolicited response. In error case
you will get solicited response or both types. To handle this we don't have work in the
mind that one command will get exactly one response with a predefined type.

The new enqueueAsyncNew method takes care about this and accepts as additional parameter a
callback function which will be called when recieving a response for this command (mapped
by the message reference id).