libmsmcomm: introduce value attribute for protocol message fields to set constant values
In some case the messages constant values which we don't want to add to the glue code
within libmsmcomm. We can now specifiy a value attribute for a field entry in the protocol
structure specification to set the constant value automatically on creation of the
structure. It simply looks like this:

<field name="foo" start="3" end="6" type="uint32_t" value="0xffff"/>

For the content of the value attribute is written as-is into the generated C code.