libmsmrpc: initial import from the android open source project
This will be the starting point for experimentaions with the variant of the msmcomm
protocol used mainly on smartphones based completly on a system made by qualcomm. Both AP
and BP processors are made by Qualcomm in this systems so they communicates over a shared
memory interface. For the communication the oncrpc protocol is used. On top of this other
protocols like AT or msmcomm (sometimes even named QMI/HCI) can be used.